refactor: move exception handlers at right place after merge
